+++ packages/adhere-mobile-ui-anthoc/es/cascader-view/index.js
@@ -1,2 +1,2 @@
-import CascaderView from"./CascaderView";export default CascaderView;
+import CascaderView from"./CascaderView";import FilterCascaderView from"./FilterCascaderView";CascaderView.FilterCascaderView=FilterCascaderView;export default CascaderView;
 //# sourceMappingURL=index.js.map