+++ packages/adhere-mobile-ui-anthoc/es/types.d.ts
@@ -1,4 +1,5 @@
-import type { CheckListProps as AntMobileCheckListProps, CheckboxProps as AntMobileCheckbox, CheckboxGroupProps as AntMobileCheckboxGroupProps, RadioGroupProps as AntMobileRadioGroupProps, RadioProps as AntMobileRadioProps, CalendarPickerViewProps, CheckListItemProps, DatePickerViewProps, DialogProps, ModalProps, PopupProps, PullToRefreshProps, SearchBarProps, SelectorOption, SelectorProps, SpaceProps } from 'antd-mobile';
+import type { CheckListProps as AntMobileCheckListProps, CheckboxProps as AntMobileCheckbox, CheckboxGroupProps as AntMobileCheckboxGroupProps, RadioGroupProps as AntMobileRadioGroupProps, RadioProps as AntMobileRadioProps, CalendarPickerViewProps, CascaderViewProps, CheckListItemProps, DatePickerViewProps, DialogProps, ModalProps, PopupProps, PullToRefreshProps, SearchBarProps, SelectorOption, SelectorProps, SpaceProps } from 'antd-mobile';
+import type { CascaderOption } from 'antd-mobile/es/components/cascader-view';
 import type { CheckListValue } from 'antd-mobile/es/components/check-list';
 import type { Action } from 'antd-mobile/es/components/modal';
 import type { CSSProperties, ReactElement, ReactNode } from 'react';
@@ -60,10 +61,10 @@ export type CheckAllWrapperProps = {
     onCheckAllChange?: (checkedValue: CheckListValue[], checked: boolean, changeValue: CheckListValue[]) => void;
     children?: ReactNode;
 };
-export type ListFilterOption<Option> = (inputValue: string, option: Option) => boolean;
+export type FilterOption<Option> = (inputValue: string, option: Option) => boolean;
 type filterProps<Option> = SearchBarProps & {
     optionFilterProp?: string;
-    filterOption?: ListFilterOption<Option> | boolean;
+    filterOption?: FilterOption<Option> | boolean;
 };
 type FilterProps<Option> = {
     filterWrapperClassName?: string;
@@ -96,7 +97,7 @@ export type ListCheckAllProps = CheckAllWrapperStyleProps & CheckAllWrapperProps
 export type ListFilterProps<Option> = {
     options: Option[];
     filterProps: filterProps<Option>;
-    children: (options: Option[]) => ReactElement;
+    children: (options: Option[], filterValue: string) => ReactElement;
     wrapperClassName?: string;
     filterWrapperClassName?: string;
     bodyWrapperClassName?: string;
@@ -104,6 +105,12 @@ export type ListFilterProps<Option> = {
     filterWrapperStyle?: CSSProperties;
     bodyWrapperStyle?: CSSProperties;
     renderEmpty?: FilterProps<Option>['renderEmpty'];
+};
+export type TreeFilterProps = Omit<ListFilterProps<any>, 'options'> & {
+    treeData?: CascaderOption[] | (Omit<CascaderOption, 'children'> & {
+        pId: string | number;
+    })[];
+    treeDataSimpleMode?: boolean;
 };
 export type CheckListHOCComponent = ReturnType<typeof createFactory<CheckListProps>> & {
     CheckAllCheckList: FC<CheckAllCheckListProps>;
@@ -418,5 +425,19 @@ export type CalendarTimestampValueHOCProps = Omit<CalendarPickerViewProps | Cale
     onChange?: (data: [number, number] | number | null | undefined) => void;
     type?: ['milliseconds' | 'seconds', 'milliseconds' | 'seconds'] | 'milliseconds' | 'seconds';
     children: ReactElement;
+};
+export type InternalCascaderViewProps = CascaderViewProps & {
+    treeDataSimpleMode?: boolean;
+};
+export type FilterCascaderViewProps = BaseType & InternalCascaderViewProps & {
+    treeDataSimpleMode?: boolean;
+    renderLabel?: (option: CascaderOption | (Omit<CascaderOption, 'children'> & {
+        pId: string | number;
+    }), filterValue: string) => ReactNode;
+} & FilterProps<CascaderOption | (Omit<CascaderOption, 'children'> & {
+    pId: string | number;
+})>;
+export type CascaderViewHOCComponent = ReturnType<typeof createFactory<InternalCascaderViewProps>> & {
+    FilterCascaderView: FC<FilterCascaderViewProps>;
 };
 export {};