+++ packages/adhere-mobile-ui-anthoc/lib/AsyncPaging.js
@@ -1,2 +1,0 @@
-"use strict";
-//# sourceMappingURL=AsyncPaging.js.map