+++ packages/adhere-mobile-ui-anthoc/lib/cascader-view/FilterCascaderView.d.ts
@@ -1,0 +1,4 @@
+import React from 'react';
+import type { DisplayNameInternal, FilterCascaderViewProps } from '../types';
+declare const FilterCascaderView: DisplayNameInternal<React.NamedExoticComponent<FilterCascaderViewProps>>;
+export default FilterCascaderView;