+++ packages/adhere-mobile-ui-anthoc/src/cascader-view/index.ts
@@ -1,3 +1,5 @@
 import CascaderView from './CascaderView';
+import FilterCascaderView from './FilterCascaderView';
 
+CascaderView.FilterCascaderView = FilterCascaderView;
 export default CascaderView;