+++ packages/adhere-mobile-ui-anthoc/changelog/CHANGELOG.md
@@ -1,3 +1,29 @@
+# 1.0.1
+
+***
+
+2024-03-06
+
+* 加入date-modal
+* 加入date-dialog
+* 加入date-popup
+* 加入time-modal
+* 加入time-dialog
+* 加入time-popup
+* 加入calendar-modal
+* 加入calendar-dialog
+* 加入calendar-popup
+* 加入ModalTrigger
+* 加入ModalTriggerPrompt
+* 加入DialogTrigger
+* 加入DialogTriggerPrompt
+* 加入PopupTrigger
+* 加入PopupTriggerPrompt
+* 对check-list进行增强
+* 对checkbox进行增强
+* 对radio进行增强
+* 对selector进行增强
+
 # 1.0.0
 
 ***